+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 * Refactored code to be a bit more clear.
 * Fixed incorrect endian-ness of the 24-bit reading and writing.
-* Changed `Header::from::<&[u8]>` to `Header::try_from::<&[u8]>` for safety.
+* Changed various `Into` and `TryInto` impls to `From` and `TryFrom` respectively.
+* Updated a `From` impl that could panic into a `TryFrom` to prevent said panic.
 
 ## Version 0.4.1